chore: Moved all sample_sketches pytest fixtures into conftext
